backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-telegram-bot==21.0.1
tweepy==4.14.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
aiofiles==24.1.0
